Refined UI look (sidebar and sample inspector), new "compact" fragment labels, improved anti-collision of fragment labels,  improved printing, bin import, bin symmetrical resizing , "apply panel" button,  marker labels no longer clip their contents, new genotype columns, column tooltips, improved visual feedback when hiding/showing rows in the detailed view, new user guide, Bug fixes: the genotype table not updating or causing crashes during some undo operation, UI issues on macOS < 11, and others.

+++ STRyper/Helpers and shared UI objects/Sorting/TableSortPopover.m
@@ -87,7 +87,7 @@
 	/// We fit the sort criteria editor its intrinsic height so that it doesn't have to scroll.
 	/// The editor itself cannot be resized (possibly due to contraints), so, we resize ourselves, which resizes the editor.
 	NSSize intrinsicContentSize = sortCriteriaEditor.intrinsicContentSize;
-	float tableHeight = intrinsicContentSize.height;
+	CGFloat tableHeight = intrinsicContentSize.height;
 	if(tableHeight < 10) {
 		return;
 	}
